+++ src/test/java/journeymap/client/api/impl/MockClientAPI.java
@@ -28,6 +28,9 @@ import journeymap.client.api.display.Context;
 import journeymap.client.api.display.DisplayType;
 import journeymap.client.api.display.Displayable;
 import journeymap.client.api.event.ClientEvent;
+import journeymap.client.api.util.MapState;
+import net.minecraft.util.AxisAlignedBB;
+import net.minecraft.util.BlockPos;
 import net.minecraftforge.fml.common.Optional;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -61,9 +64,17 @@ enum MockClientAPI implements journeymap.client.api.IClientAPI
 
 
     @Override
-    public boolean isActive(Enum<? extends Context>... contexts)
+    public MapState getMapState(Context.UI uiContext)
     {
-        return true;
+        if (uiContext == Context.UI.Any)
+        {
+            return null;
+        }
+
+        return new MapState(Context.UI.Fullscreen, true, 0, 1,
+                Context.MapType.Day,
+                new BlockPos(128, 0, 128),
+                new AxisAlignedBB(new BlockPos(0, 0, 0), new BlockPos(256, 256, 256)));
     }
 
     @Override